Things working a bit better now.
All the fields were getting overwritten other than the one you just
typed into. Now it's actually submitting the form so all the form values
show up in the URL instead of only the one you just changed.

That morph stuff makes it so you don't lose keyboard focus even when
things change out. Magical. A shame it's another file to include, but
it's gonna be part of the main htmx soon apparently.

https://htmx.org/docs/#morphing